restructure init and system config
